Move files for config
