+++ test/config/copy.bpmn.diagram.ts
@@ -13,5 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { toMatchImageSnapshot } from 'jest-image-snapshot';
-expect.extend({ toMatchImageSnapshot });
+import { copySync } from 'fs-extra';
+
+function copyDirectoryContentIfNotExists(dirName: string): void {
+  copySync(`${__dirname}/../fixtures/bpmn/${dirName}`, `${__dirname}/../../build/public/static/diagrams/`, { overwrite: true });
+}
+
+copyDirectoryContentIfNotExists('diagram');
+copyDirectoryContentIfNotExists('navigation');
+copyDirectoryContentIfNotExists('non-regression');
+copyDirectoryContentIfNotExists('overlays');
+copyDirectoryContentIfNotExists('performance');
+copyDirectoryContentIfNotExists('svg');
+copyDirectoryContentIfNotExists('theme');
