+++ test/config/jest-playwright.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const { isRunningOnCISlowOS } = require('../environment-utils');
+const { isRunningOnCISlowOS } = require('../helpers/environment-utils');
 
 const log = require('debug')('bv:test:config:pw');
