+++ test/e2e/jest-playwright.config.js
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-module.exports = require('../helpers/config/jest-playwright').computeConfigurationForDevServerUsage();
+module.exports = require('../config/jest-playwright').computeConfigurationForDevServerUsage();
