+++ test/e2e/jest.config.js
@@ -34,13 +34,13 @@ module.exports = {
   coveragePathIgnorePatterns: ['/src/model'],
   coverageReporters: ['lcov', 'text-summary'],
   coverageDirectory: 'build/test-report/e2e',
-  setupFiles: ['./test/e2e/config/copy.bpmn.diagram.ts'],
+  setupFiles: ['./test/config/copy.bpmn.diagram.ts'],
   setupFilesAfterEnv: [
     'expect-playwright',
     // jest-image-snapshot configuration doesn't work with setupFiles, fix with setupFilesAfterEnv: see https://github.com/testing-library/jest-dom/issues/122#issuecomment-650520461
-    './test/e2e/config/jest.image.ts',
+    './test/config/jest.image.ts',
     // need playwright globals to be available, so after environment
-    './test/e2e/config/playwright.ts',
+    './test/config/playwright.ts',
   ],
   reporters: [
     'default',
