+++ test/performance/jest.config.js
@@ -30,6 +30,6 @@ module.exports = {
       tsconfig: './tsconfig.json',
     },
   },
-  setupFiles: ['./test/e2e/config/copy.bpmn.diagram.ts'],
+  setupFiles: ['./test/config/copy.bpmn.diagram.ts'],
   setupFilesAfterEnv: ['expect-playwright'],
 };
